% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/murdock.R
\docType{data}
\name{murdock}
\alias{murdock}
\title{Murdock Ethnographic Atlas}
\format{An object of class \code{SpatialPolygonsDataFrame} with 844 rows and 124 columns.}
\usage{
murdock
}
\description{
Murdock Ethnographic Atlas
}
\keyword{datasets}

